changing to true color and adding some gu/gum 3D functions for testing

+++ LuaPlayer/src/framebuffer.c
@@ -1,3 +1,3 @@
 #include <psptypes.h>
 
-u16* g_vram_base = (u16*) (0x40000000 | 0x04000000);
+u32* g_vram_base = (u16*) (0x40000000 | 0x04000000);
